test: modularize, rename and comment
Refator the integration test to make room for client/server test.

The integration test is renamed to reflect its testing the offline mode
of the program. The circuit definition is moved to another file, so it's
clearly separated from the test logic. It will be later used by the
online mode test. Separate test resources from code.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>