randomness: make testable, mockable and add tests
Refactor the package so it has a settable provider. Convert the existing
init function to the real drand-based provided. Add another mock
provider that returns a fixed value instead of a proper random one.

Override the drand provided with the mock provider in tests. This way
tests can run and pass when e.g. the service is dead or when I'm on the
plane again with no internet access.

Move it to utils because this is where all the goodies that are helpful
but not really related to the trusted setup itself are going to live.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>