phase2: accept beacon as argument
The randomness package generates beacon once at init. This does not make
sense in the offline mode, where the application is ran multiple times.
Such approach will generate a different random value at each run, while
it must be the same at init and at extract-keys.

Change the approach. The extract-keys command now accepts a --beacon
argument. User is expect to pass the same beacon that was generated and
printed on the init command.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>