online: server: store ceremony log to file
Ceremony log is necessary for ceremony atestation. It documents
the course of contributions and contains beacon, which is necessary for
regenerating keys out of original contributions in the offline mode.

The log is stored in persistent storage in a structured JSON format. It
is also pretty-printed to stdout during the ceremony.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>